+++ tests/testCompPubKey.py
@@ -29,3 +29,21 @@ class OpensslTest(TestCase):
         publicKey = PublicKey.fromCompressed(compPubKeyString)
 
         self.assertTrue(publicKeyString == publicKey.toString())
+
+    def compareToCompressedEven(self):
+        privateKey = PrivateKey.fromString("f91d8f3a49805fff9289769247e984b355939679f3080156fe295229e00f25af")
+        publicKey = privateKey.publicKey()
+
+        compPublicKey = publicKey.toCompressed()
+        controlPublicKey = "0252972572d465d016d4c501887b8df303eee3ed602c056b1eb09260dfa0da0ab2"
+
+        self.assertTrue(compPublicKey == controlPublicKey)
+
+    def compareToCompressedOdd(self):
+        privateKey = PrivateKey.fromString("ac609e0cc9681f8cb63e968be20e0f19721751561944f5b4e52d54d5f27ec57b")
+        publicKey = privateKey.publicKey()
+
+        compPublicKey = publicKey.toCompressed()
+        controlPublicKey = "0318ed2e1ec629e2d3dae7be1103d4f911c24e0c80e70038f5eb5548245c475f50"
+
+        self.assertTrue(compPublicKey == controlPublicKey)